Add support for MongoDB and custom OTP model
This commit adds support for using the package with MongoDB by creating
a new Otp model and using it in the AppServiceProvider provider. It also
adds the ability to use a custom OTP model. The commit also includes
updates to the Otp class to set the 'valid' attribute to true when
creating a new OTP.